Maybe fix some test issues
